Cleanup

+++ contracts/interfaces/IWETH.sol
@@ -2,8 +2,12 @@ pragma solidity >=0.5.0;
 
 interface IWETH {
     function deposit() external payable;
-    function transfer(address to, uint value) external returns (bool);
-    function withdraw(uint) external;
-    function balanceOf(address guy) external returns (uint);
-    function approve(address guy, uint wad) external returns (bool);
+
+    function transfer(address to, uint256 value) external returns (bool);
+
+    function withdraw(uint256) external;
+
+    function balanceOf(address guy) external returns (uint256);
+
+    function approve(address guy, uint256 wad) external returns (bool);
 }
