+++ contracts/libraries/DXswapOracleLibrary.sol
@@ -1,8 +1,8 @@
 pragma solidity >=0.5.0;
 
-import 'dxswap-core/contracts/interfaces/IDXswapPair.sol';
+import "dxswap-core/contracts/interfaces/IDXswapPair.sol";
 
-import './FixedPoint.sol';
+import "./FixedPoint.sol";
 
 // library with helper methods for oracles that are concerned with computing average prices
 library DXswapOracleLibrary {
@@ -10,13 +10,19 @@ library DXswapOracleLibrary {
 
     // helper function that returns the current block timestamp within the range of uint32, i.e. [0, 2**32 - 1]
     function currentBlockTimestamp() internal view returns (uint32) {
-        return uint32(block.timestamp % 2 ** 32);
+        return uint32(block.timestamp % 2**32);
     }
 
     // produces the cumulative price using counterfactuals to save gas and avoid a call to sync.
-    function currentCumulativePrices(
-        address pair
-    ) internal view returns (uint price0Cumulative, uint price1Cumulative, uint32 blockTimestamp) {
+    function currentCumulativePrices(address pair)
+        internal
+        view
+        returns (
+            uint256 price0Cumulative,
+            uint256 price1Cumulative,
+            uint32 blockTimestamp
+        )
+    {
         blockTimestamp = currentBlockTimestamp();
         price0Cumulative = IDXswapPair(pair).price0CumulativeLast();
         price1Cumulative = IDXswapPair(pair).price1CumulativeLast();
@@ -28,9 +34,9 @@ library DXswapOracleLibrary {
             uint32 timeElapsed = blockTimestamp - blockTimestampLast;
             // addition overflow is desired
             // counterfactual
-            price0Cumulative += uint(FixedPoint.fraction(reserve1, reserve0)._x) * timeElapsed;
+            price0Cumulative += uint256(FixedPoint.fraction(reserve1, reserve0)._x) * timeElapsed;
             // counterfactual
-            price1Cumulative += uint(FixedPoint.fraction(reserve0, reserve1)._x) * timeElapsed;
+            price1Cumulative += uint256(FixedPoint.fraction(reserve0, reserve1)._x) * timeElapsed;
         }
     }
 }
